+++ packages/isar/lib/src/native/split_words.dart
@@ -1,9 +1,6 @@
 import 'package:isar/isar.dart';
 import 'package:isar/src/native/isar_core.dart';
 
-import 'binary_reader.dart';
-import 'binary_writer.dart';
-
 List<String> splitWordsNative(String input) {
   initializeIsarCore();