+++ packages/isar_test/lib/float_model.dart
@@ -1,5 +1,7 @@
 import 'package:isar/isar.dart';
 
+import 'common.dart';
+
 part 'float_model.g.dart';
 
 @Collection()
@@ -14,15 +16,11 @@ class FloatModel {
   @Size32()
   List<double>? list;
 
-  @Index(type: IndexType.hash)
-  @Size32()
-  List<double>? hashList;
-
   FloatModel();
 
   @override
   String toString() {
-    return '{field: $field, list: $list, hashList: $hashList}';
+    return '{field: $field, list: $list}';
   }
 
   @override
@@ -32,26 +30,8 @@ class FloatModel {
       return false;
     } else if (field != null && (otherModel.field! - field!).abs() > 0.001) {
       return false;
-    } else if (other.list?.length != list?.length) {
+    } else if (!doubleListEquals(list, other.list)) {
       return false;
-    } else if (other.hashList?.length != hashList?.length) {
-      return false;
-    }
-
-    if (list != null) {
-      for (var i = 0; i < list!.length; i++) {
-        if ((otherModel.list![i] - list![i]).abs() > 0.001) {
-          return false;
-        }
-      }
-    }
-
-    if (hashList != null) {
-      for (var i = 0; i < hashList!.length; i++) {
-        if ((otherModel.hashList![i] - hashList![i]).abs() > 0.001) {
-          return false;
-        }
-      }
     }
 
     return true;